Experimental package build (#847)
