+++ cli/dcoscli/data/help/package.txt
@@ -5,7 +5,6 @@ Usage:
     dcos package --config-schema
     dcos package --help
     dcos package --info
-    dcos package build [--output-directory=<output-directory>] <build-definition>
     dcos package describe [--app --cli --config]
                           [--render]
                           [--package-versions]
@@ -27,8 +26,6 @@ Usage:
     dcos package update
 
 Commands:
-    build
-        Build a package to install to DC/OS or share with Universe.
     describe
         Get specific details for packages.
     install
@@ -71,9 +68,6 @@ Options:
         Print a short description of this subcommand.
     --options=<file>
         Path to a JSON file that contains customized package installation options.
-    --output-directory=<output-directory>
-        Path to the directory where the data should be stored.
-        Defaults to the current working directory.
     --package-version=<package-version>
         The package version to install.
     --package-versions
@@ -87,8 +81,6 @@ Options:
         Turn off interactive mode and assume "yes" is the answer to all prompts.
 
 Positional Arguments:
-    <build-definition>
-        Path to a DC/OS Package Build Definition.
     <package-name>
         Name of the DC/OS package in the package repository.
     <query>
